Revamp, dedup and refactor techinical indicators
