+++ algorithms/heikin_ashi_spike_hunter.py
@@ -10,11 +10,8 @@ from algorithms.heikin_ashi import HeikinAshi
 from models.signals import BollinguerSpread, SignalsConsumer
 from shared.enums import Strategy
 
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
-
 if TYPE_CHECKING:
-    from binquant.producers.technical_indicators import TechnicalIndicators
+    from producers.analytics import CryptoAnalytics
 
 
 class HASpikeHunter(HeikinAshi):
@@ -32,7 +29,7 @@ class HASpikeHunter(HeikinAshi):
 
     def __init__(
         self,
-        cls: "TechnicalIndicators",
+        cls: "CryptoAnalytics",
     ):
         script_dir = path.dirname(__file__)
         rel_path = "checkpoints/spikehunter_model_v1.pkl"
@@ -178,9 +175,9 @@ class HASpikeHunter(HeikinAshi):
             signal = 0
             signal_type = ""
             strength = 0.0
-            current_price_change = df.loc[i, "price_change"]
-            current_volume_ratio = df.loc[i, "volume_ratio"]
-            current_rsi = df.loc[i, "rsi"]
+            current_price_change = df.iloc[i]["price_change"]
+            current_volume_ratio = df.iloc[i]["volume_ratio"]
+            current_rsi = df.iloc[i]["rsi"]
             # Method 1: ML classifier
             if ml_preds[i] == 1:
                 signal = 1
@@ -196,7 +193,7 @@ class HASpikeHunter(HeikinAshi):
                 strength = min(current_price_change * current_volume_ratio * 10, 10.0)
             # Method 3: Momentum Detection
             elif i >= 5:
-                recent_changes = df.loc[i - 3 : i, "price_change"]
+                recent_changes = df.iloc[i - 3 : i]["price_change"]
                 positive_moves = (recent_changes > self.momentum_threshold).sum()
                 total_momentum = recent_changes.sum()
                 if positive_moves >= 2 and total_momentum > price_thresh:
@@ -207,7 +204,7 @@ class HASpikeHunter(HeikinAshi):
             elif (
                 i >= 14
                 and current_rsi > self.rsi_oversold
-                and df.loc[i - 1, "rsi"] <= self.rsi_oversold
+                and df.iloc[i - 1]["rsi"] <= self.rsi_oversold
                 and current_price_change > 0.008
             ):
                 signal = 1
@@ -216,9 +213,9 @@ class HASpikeHunter(HeikinAshi):
                     (current_rsi - self.rsi_oversold) / 10 * current_price_change * 50,
                     6.0,
                 )
-            df.loc[i, "spike_signal"] = signal
-            df.loc[i, "spike_type"] = signal_type
-            df.loc[i, "signal_strength"] = strength
+            df.at[i, "spike_signal"] = signal
+            df.at[i, "spike_type"] = signal_type
+            df.at[i, "signal_strength"] = strength
         return df
 
     def get_spike_summary(self, df: pd.DataFrame) -> dict:
