+++ algorithms/market_breadth.py
@@ -6,11 +6,11 @@ from models.signals import BollinguerSpread, SignalsConsumer
 from shared.enums import MarketDominance, Strategy
 
 if TYPE_CHECKING:
-    from producers.technical_indicators import TechnicalIndicators
+    from producers.analytics import CryptoAnalytics
 
 
 class MarketBreadthAlgo:
-    def __init__(self, cls: "TechnicalIndicators") -> None:
+    def __init__(self, cls: "CryptoAnalytics") -> None:
         self.ti = cls
         self.current_market_dominance = MarketDominance.NEUTRAL
         self.btc_change_perc = 0
